+++ src/taxify/Taxi.java
@@ -1,14 +1,14 @@
 package taxify;
 
 public class Taxi extends Vehicle {
-    private static final int RATE = 2;
+    private static final double RATE = 2;
 
     public Taxi(int id, ILocation location) {
         super(id, location);
     }
 
     @Override
-    public int calculateCost() {
+    public double calculateCost() {
         return super.calculateCost() * RATE;
     }
 }
